backend/pyproject.toml
[project]
name = "Jarcode"
version = "0.1.0"
description = "API for Jarcode application."
requires-python = ">=3.10"
dependencies = [
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.1",
    "django==5.2.0",
    "django-cleanup>=9.0.0",
    "django-dramatiq>=0.14.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework==3.16.1",
    "docker>=7.1.0",
    "dramatiq>=1.18.0",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "google-genai>=1.53.0",
    "nh3>=0.3.2",
    "pika>=1.3.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "rabbitmq>=0.2.0",
    "redis>=6.4.0",
    "requests>=2.32.5",
]
